+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/EResourcesDetails.test.js
@@ -23,9 +23,8 @@ jest.mock('@folio/stripes/components', () => ({
 
 jest.mock('../../../hooks', () => ({
   useFieldMappingBoolFieldValue: () => ['ALL_FALSE', 'ALL_FALSE'],
-  useFieldMappingFieldValue: () => ['testPOStatus'],
+  useFieldMappingFieldValue: () => ['Open'],
   useFieldMappingValueFromLookup: () => ['testUUID', 'testMapping'],
-  useFieldMappingFieldValueByPath: () => ['"None"'],
 }));
 
 const okapiProp = {