+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/PhysicalResourceDetails.test.js
@@ -20,9 +20,8 @@ jest.mock('@folio/stripes/components', () => ({
 }));
 
 jest.mock('../../../hooks', () => ({
-  useFieldMappingFieldValue: () => ['Open'],
+  useFieldMappingFieldValue: () => ['Open', 'Electronic resource'],
   useFieldMappingValueFromLookup: () => ['', ''],
-  useFieldMappingFieldValueByPath: () => ['"None"'],
   useFieldMappingRefValues: () => [[{
     order: 0,
     path: 'order.poLine.physical.volumes[]',