+++ src/components/MARCTableView/MARCTableViewRowContainer.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../test/jest/__mock__';
@@ -77,6 +78,12 @@ const renderMARCTableViewRowContainer = ({
 };
 
 describe('MARCTableViewRowContainer', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderMARCTableViewRowContainer(MARCTableViewRowContainerProps);
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should be rendered with correct cells', () => {
     const { getAllByText } = renderMARCTableViewRowContainer(MARCTableViewRowContainerProps);