+++ src/components/MatchCriterion/view/QualifierPartSection/QualifierPartSection.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../../test/jest/__mock__';
@@ -38,6 +39,12 @@ const renderQualifierPartSection = ({
 };
 
 describe('QualifierPartSection view', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderQualifierPartSection(qualifierPartSection);
+
+    await runAxeTest({ rootNode: container });
+  });
+
   describe('when there is a value', () => {
     it('should be rendered with additional content', () => {
       const { container } = renderQualifierPartSection(qualifierPartSection);