+++ src/components/MatchCriterion/view/ViewMatchCriterion.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../test/jest/__mock__';
@@ -68,6 +69,12 @@ const renderViewMatchCriterion = ({
 };
 
 describe('ViewMatchCriterion', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderViewMatchCriterion({ ...viewMatchCriterionProps('EXACTLY_MATCHES') });
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should be rendered with a qualifier type', () => {
     const { getByText } = renderViewMatchCriterion({ ...viewMatchCriterionProps('EXACTLY_MATCHES') });