+++ src/components/TreeView/TreeView.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-
 import { screen } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import '../../../test/jest/__mock__';
 
@@ -70,6 +70,12 @@ const renderTreeView = treeViewProps => {
 };
 
 describe('Tree View component', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderTreeView(mockSpacingProps);
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('Should be rendered', () => {
     expect(renderTreeView(mockSpacingProps)).toBeDefined();
   });