+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/tests/InvoiceLineInformation.test.js
@@ -5,6 +5,7 @@ import {
 } from '@testing-library/react';
 import { get } from 'lodash';
 
+import { runAxeTest } from '@folio/stripes-testing';
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
 import '../../../../../../../test/jest/__mock__';
@@ -42,6 +43,12 @@ const renderInvoiceLineInformation = () => {
 describe('InvoiceLineInformation', () => {
   afterAll(() => {
     setReferenceTablesMock.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderInvoiceLineInformation();
+
+    await runAxeTest({ rootNode: container });
   });
 
   it('should render correct fields', () => {