+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/CostDetails.test.js
@@ -4,6 +4,7 @@ import {
   within,
 } from '@testing-library/react';
 import { noop } from 'lodash';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
@@ -37,6 +38,12 @@ const renderCostDetails = () => {
 };
 
 describe('CostDetails', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderCostDetails();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render correct fields', async () => {
     const { getByText } = renderCostDetails();