+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/EResourcesDetails.test.js
@@ -3,6 +3,7 @@ import {
   fireEvent,
   within,
 } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
@@ -49,6 +50,12 @@ const renderEResourcesDetails = () => {
 };
 
 describe('EResourcesDetails', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderEResourcesDetails();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render correct fields', async () => {
     const { getByText } = renderEResourcesDetails();