+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/OrderInformation.test.js
@@ -3,6 +3,7 @@ import {
   fireEvent,
   within,
 } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import { buildResources } from '@folio/stripes-data-transfer-components/test/helpers';
@@ -104,6 +105,12 @@ const renderOrderInformation = () => {
 describe('OrderInformation', () => {
   afterEach(() => {
     setReferenceTablesMock.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderOrderInformation();
+
+    await runAxeTest({ rootNode: container });
   });
 
   it('should render correct fields', async () => {