+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bugs fixed:
 * Order field mapping profile: Fix the UI and error message for the fund "Value" field (UIDATIMP-1387)
+* Order field mapping profile: Add info icon to the Purchase order status fieldPurchase order status a required field (UIDATIMP-1397)
 
 ## [6.0.1](https://github.com/folio-org/ui-data-import/tree/v6.0.1) (2023-03-14)